refactor(riscv): flatten width-aligned load store modules
