+++ extensions/riscv/circuit/src/load/core.rs
