--- extensions/riscv/circuit/src/load_sign_extend/width_aligned/mod.rs
@@ -1,1 +1,0 @@
-pub mod core;
