+++ extensions/riscv/circuit/src/store/core.rs
