+++ extensions/riscv/circuit/src/store/doubleword/mod.rs
@@ -7,7 +7,7 @@ use crate::{
     },
     store::{
         common::StoreExecutor,
-        width_aligned::{StoreWidthAlignedCoreAir, StoreWidthAlignedFiller},
+        core::{StoreWidthAlignedCoreAir, StoreWidthAlignedFiller},
     },
 };
