--- extensions/riscv/circuit/src/store/width_aligned/mod.rs
@@ -1,2 +1,0 @@
-pub mod core;
-pub use core::*;
