+++ extensions/riscv/circuit/src/test_utils/memory.rs
@@ -58,17 +58,17 @@ pub(crate) use crate::{
         Rv64StoreAdapterCols, Rv64StoreAdapterExecutor, Rv64StoreAdapterFiller, RV64_BYTE_BITS,
     },
     load::{
-        common::load_write_data, width_aligned::LoadWidthAlignedCoreCols, LoadByteCoreAir,
-        LoadByteCoreCols, LoadByteFiller, LoadDoublewordCoreAir, LoadDoublewordFiller,
-        LoadHalfwordCoreAir, LoadHalfwordFiller, LoadWordCoreAir, LoadWordFiller, Rv64LoadByteAir,
-        Rv64LoadByteChip, Rv64LoadByteExecutor, Rv64LoadDoublewordAir, Rv64LoadDoublewordChip,
+        common::load_write_data, core::LoadWidthAlignedCoreCols, LoadByteCoreAir, LoadByteCoreCols,
+        LoadByteFiller, LoadDoublewordCoreAir, LoadDoublewordFiller, LoadHalfwordCoreAir,
+        LoadHalfwordFiller, LoadWordCoreAir, LoadWordFiller, Rv64LoadByteAir, Rv64LoadByteChip,
+        Rv64LoadByteExecutor, Rv64LoadDoublewordAir, Rv64LoadDoublewordChip,
         Rv64LoadDoublewordExecutor, Rv64LoadHalfwordAir, Rv64LoadHalfwordChip,
         Rv64LoadHalfwordExecutor, Rv64LoadWordAir, Rv64LoadWordChip, Rv64LoadWordExecutor,
         DOUBLEWORD_LOAD_SELECTOR_WIDTH, HALFWORD_LOAD_SELECTOR_WIDTH, WORD_LOAD_SELECTOR_WIDTH,
     },
     load_sign_extend::common::load_sign_extend_write_data,
     store::{
-        common::store_write_data, width_aligned::StoreWidthAlignedCoreCols, Rv64StoreByteAir,
+        common::store_write_data, core::StoreWidthAlignedCoreCols, Rv64StoreByteAir,
         Rv64StoreByteChip, Rv64StoreByteExecutor, Rv64StoreDoublewordAir, Rv64StoreDoublewordChip,
         Rv64StoreDoublewordExecutor, Rv64StoreHalfwordAir, Rv64StoreHalfwordChip,
         Rv64StoreHalfwordExecutor, Rv64StoreWordAir, Rv64StoreWordChip, Rv64StoreWordExecutor,
